various stuff
- changed glass mentioned audio datablock name to glass bell, not
restricted just as a mentioned sound.
- adding friend now accepts friend request if there is already one in
user info window.
- fixed bl name changing not updating client-side.
- pending friend requests on glass live connect notification (comes with
option to turn off, on by default).
- updatesgroup startup error fix.
- glassliveuser::getfromname(%name); added
- laid down work for the user blocking system, whenever that'll come.

+++ client/GlassUpdaterSupport.cs
@@ -148,7 +148,8 @@ function GlassUpdaterSupport::updateSetting() {
 
 package GlassUpdaterSupportPackage {
   function updater::checkForUpdates(%this, %a, %b, %c) {
-    GlassUpdatesGroup.delete();
+    if(isObject(GlassUpdatesGroup))
+      GlassUpdatesGroup.delete();
     return parent::checkForUpdates(%this, %a, %b, %c);
   }
